Allow customization of the "quote_char"
Currently, csvview.nvim uses both " and ' as quote chars.
This causes problems with some files, where for instance the ' char is
used in the language (e.g. "It's a nice day to be alive!").

Change the behavior of csvview.nvim to use the " as quote char by
default, but allows the user to specify another quote char in the
options (both in the configuration and as a parameter to the CsvView...
commands).