feat(view): Support sticky header (#39)
* refactor(view): extract extmark setting logic into _add_extmark method

* feat: improve multi-window handling in csvview

* feat(view): add sticky header feature

Added a sticky header feature to the CSV view. The header line
remains visible at the top of the window when scrolling. The
feature can be enabled or disabled via the `sticky_header`
option in the view configuration. The `header_lnum` option
specifies the line number of the header, which is used for the
sticky header feature.

* feat: Add highlight for sticky-header separator

* docs: update README

* test: add test for sticky-header feature

* tests: fix test

* fix(sticky_header): add compatibility for Neovim versions below 0.11

* fix(sticky-header): fix sticky-header highlight